backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Main dependencies - required for production
fastapi
uvicorn[standard]
supabase
python-multipart

# Explicit core numeric dependency — required by easyocr, opencv, and torch.
# Declaring it here ensures pip installs it before the packages that need it,
# preventing "Numpy is not available" errors in constrained environments.
numpy<2.0.0

# OCR - CPU-only slim builds to minimise image size and runtime memory.
# Pinning to the +cpu variants ensures pip never pulls the full CUDA wheels
# (~2 GB each) even when the PyPI index is consulted first.
torch==2.2.2+cpu
torchvision==0.17.2+cpu
easyocr
opencv-python-headless
pillow
pdf2image

# Database
asyncpg

# Auth & API
python-jose[cryptography]
httpx
pydantic-settings
python-dotenv
bcrypt
